requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
filelock==3.7.1
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
numpy==1.21.6
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.4
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
